Move class to the correct package
